requirements.txt
streamlit
pillow
pyzbar
openpyxl